repo clean up andrei suggestion
