+++ .gitignore
@@ -3,6 +3,8 @@ __pycache__/
 *.py[cod]
 *$py.class
 
+# versions
+_version.py
 # C extensions
 *.so
