+++ old_tavi/convert_spice_to_nexus.py
