+++ old_tavi/data/spice_reader.py
