+++ old_tavi/instrument/components/collimators.py
