+++ old_tavi/instrument/components/component_base.py
