+++ old_tavi/instrument/components/detector.py
