+++ old_tavi/instrument/components/monitor.py
