+++ old_tavi/instrument/components/mono_ana.py
