+++ old_tavi/instrument/instrument_params/cg4c.json
