+++ old_tavi/instrument/instrument_params/hb3.json
