+++ old_tavi/instrument/resolution/cooper_nathans.py
