+++ old_tavi/instrument/resolution/curve.py
