+++ old_tavi/instrument/resolution/ellipse.py
