+++ old_tavi/instrument/resolution/ellipsoid.py
