+++ old_tavi/instrument/resolution/resolution_calculator.py
