+++ old_tavi/lattice_algorithm.py
