+++ old_tavi/old_tests/test_spice_reader.py
