+++ old_tavi/sample.py
