+++ old_tavi/tavimain.py
@@ -6,7 +6,7 @@ import sys
 
 from qtpy.QtWidgets import QApplication, QMainWindow
 
-from tavi import __version__
+from old_tavi import __version__
 from tavi.configuration import Configuration
 from tavi.mainwindow import MainWindow
